Finally normalized to backgrounds before separating subpools!!!
Was it necessary to work on all day?
Not sure

But I did it

Function substitutes shorter names for backgrounds, defines background expression then joins these to the original dataset to normalize and take the average of normalized values. Subpool separation is modified to remove these steps. Could not perform this function with controls, left separate. Updated applicable plots that reference these datasets. Also output the normalized to background, both subpool datasets to a table to compare with integrated.